fix loader list overflow bug

+++ src/function/string/operations/include/left_operation.h
@@ -19,7 +19,8 @@ public:
     static inline void operation(gf_string_t& left, int64_t& right, gf_string_t& result,
         bool isLeftNull, bool isRightNull, ValueVector& resultValueVector) {
         assert(!isLeftNull && !isRightNull);
-        auto len = right > 0 ? min(left.len, (uint32_t)right) : max(left.len + right, 0l);
+        auto len = right > 0 ? min(left.len, (uint32_t)right) :
+                               max(left.len + (uint32_t)right, (uint32_t)0u);
         SubStr::operation(left, 1, len, result, false /* isLeftNull*/, false /* isStartNull*/,
             false /*isLenNull */, resultValueVector);
     }
