+++ src/function/string/operations/include/right_operation.h
@@ -19,7 +19,8 @@ public:
     static inline void operation(gf_string_t& left, int64_t& right, gf_string_t& result,
         bool isLeftNull, bool isRightNull, ValueVector& resultValueVector) {
         assert(!isLeftNull && !isRightNull);
-        auto len = right > 0 ? min(left.len, (uint32_t)right) : max(0l, left.len + right);
+        auto len = right > 0 ? min(left.len, (uint32_t)right) :
+                               max((uint32_t)0u, left.len + (uint32_t)right);
         SubStr::operation(left, left.len - len + 1, len, result, false /* isLeftNull*/,
             false /* isStartNull*/, false /*isLenNull */, resultValueVector);
     }
